Display relavent links in sidebar according to Layout userRole prop
